requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
celery==5.3.4
redis==5.0.1
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
uuid==1.30
python-dateutil==2.8.2
gunicorn==21.2.0
supervisor==4.2.5
psutil==5.9.6 